improved LoadDLL
+ improved GetArgNumberFromFunction: now it can guess how many
parameters a cdecl has
+ added new column in the exported listview: now it will guess the
function type (cdecl, stdcall, etc)
+ fixed LDE64.h: oh my... on 32bit LDE was using DWORD64 as return value
instead of DWORD
+ GUI: added chkPause button

+++ LoadDll/disasm/LDE64.h
@@ -2,7 +2,7 @@
 #ifdef _WIN64
 extern "C" { DWORD64 _fastcall LDE(DWORD64 dwAddr, int nSize); }
 #else
-extern "C" { DWORD64 _stdcall LDE(DWORD dwAddr, int nSize); }
+extern "C" { DWORD _stdcall LDE(DWORD dwAddr, int nSize); }
 #endif
